+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Yarn now accepts sponsorships! Please give a look at our [OpenCollective](https:
 - Yarn is now able to migrate classic lockfiles containing unconventional tarball URLs
 - The nm linker hoists portals after hoisting their dependencies first
 - The PnP filesystem now handles `read` and `readSync` using options
+- The PnP filesystem now handles UNC paths using forward slashes
 
 ### Miscellaneous Features